Working on simplifying plotting code - see #12
Re-implementing the bubble chart data structure; trying to move away from an absurd no. of dictionaries and global variable vomit.
All copied code should be able to be contained in a DataAndThings class of some kind.
Will be migrating the actual plotting functionality over soon, as well as the rest of the data things.